Fix for StringMatchesMask, proper handle the "*text*" mask
Function StringMatchesMask was unable to properly search for "*text*" in "This is a test text." string. in both cases where case sensitive was used or not, it found first occurence of mask letters in string, but it was not all the letters, eg: "te" was found, it returned False and not searchin to the end of string for any other matches.